fix: normalize class routine foreign keys

+++ EduOS.Core/Entities/Academic/ClassRoutine.cs
@@ -5,11 +5,11 @@ namespace EduOS.Core.Entities.Academic
 {
     public class ClassRoutine : BaseTenantEntity
     {
-        public int AcademicYearId { get; set; }
-        public int ClassId { get; set; }
-        public int SectionId { get; set; }
-        public int SubjectId { get; set; }
-        public int TeacherId { get; set; }
+        public long AcademicYearId { get; set; }
+        public long ClassId { get; set; }
+        public long SectionId { get; set; }
+        public long SubjectId { get; set; }
+        public long TeacherId { get; set; }
         public string DayOfWeek { get; set; } = string.Empty; // Saturday-Friday
         public TimeSpan StartTime { get; set; }
         public TimeSpan EndTime { get; set; }
